requirements.txt
torch
torchvision
streamlit
pillow
numpy
scikit-learn
tqdm
